we changed the whole project into a better structure
