+++ .gitignore
@@ -1,22 +1,1 @@
-# See https://help.github.com/ignore-files/ for more about ignoring files.
-
-# dependencies
 node_modules
-
-# testing
-coverage
-
-# production
-build
-
-# misc
-.DS_Store
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
-.graphcoolrc
-
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
