+++ package.json
@@ -2,23 +2,21 @@
   "name": "server",
   "version": "1.0.0",
   "description": "",
-  "main": "src/index.js",
+  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "async-request": "^1.2.0",
     "bcryptjs": "^2.4.3",
     "express-jwt": "^5.3.1",
+    "google-protobuf": "^3.8.0-rc.1",
     "graphql-yoga": "^1.17.4",
+    "grpc": "^1.20.3",
     "jsonwebtoken": "^8.5.0",
     "math": "0.0.3",
-    "mockjs": "^1.0.1-beta3",
-    "nodemon": "^1.18.10",
-    "prisma-client-lib": "^1.28.0",
-    "request": "^2.88.0"
+    "mysql": "^2.16.0",
+    "prisma-client-lib": "^1.29.1"
   }
 }
