+++ resolvers/index.js
@@ -1,4 +1,4 @@
-const { query } = require('./query')
+const { query } = require('./query/query')
 const { auth } = require('./mutation/auth')
 const { order } = require('./mutation/order')
