+++ resolvers/mutation/auth.js
@@ -1,6 +1,6 @@
 const bcrypt = require('bcryptjs')
 const jwt = require('jsonwebtoken')
-const { getUserId, getOpenId } = require('../../utils')
+const { getUserId, getOpenId } = require('../../utils/utils')
 
 const auth = {
   async login(parent, args, ctx, info) {
