+++ utils/wx_openid.js
